Fixed re-rendering issue
