+++ src/redux/rockets/rockets.js
@@ -1,5 +1,6 @@
 // Actions
 const SAVE = 'spacetravelershub/rockets/SAVE';
+// const RESERVE = 'spacetravelershub/rockets/RESERVE';
 
 // Reducer
 export default function reducer(state = [], action) {
@@ -17,6 +18,7 @@ export function saveRockets(array) {
   return () => ({
     type: SAVE,
     array: array.map((rocket) => ({
+      reserved: false,
       id: rocket.id,
       name: rocket.rocket_name,
       description: rocket.description,
